+++ packages/terra-core-docs/src/terra-dev-site/doc/action-header/example/BackPreviousNextCloseActionHeader.jsx
@@ -6,11 +6,14 @@ const ActionHeaderExample = () => (
   <div>
     <br />
     <ActionHeader
-      title="Back Close Previous Next Action Header"
+      text="Back Close Previous Next Action Header"
       onClose={() => alert('You clicked close!')}
       onBack={() => alert('You clicked back!')}
+      backButtonA11yLabel="Show Back Alert"
       onPrevious={() => alert('You clicked previous!')}
+      prevButtonA11yLabel=" Show Previous Alert"
       onNext={() => alert('You clicked next!')}
+      nextButtonA11yLabel="Show Next Alert"
     />
     <br />
   </div>