+++ packages/terra-core-docs/src/terra-dev-site/doc/action-header/example/CloseActionHeader.jsx
@@ -6,7 +6,7 @@ const ActionHeaderExample = () => (
   <div>
     <br />
     <ActionHeader
-      title="Close Action Header"
+      text="Close Action Header"
       onClose={() => alert('You clicked close!')}
     />
     <br />