+++ packages/terra-core-docs/src/terra-dev-site/doc/action-header/example/LongWrappingTextHeader.jsx
@@ -11,7 +11,7 @@ const ActionHeaderExample = () => (
   <div>
     <br />
     <ActionHeader
-      title="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla orci dolor, dignissim vitae risus vel, tristique egestas sapien.
+      text="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla orci dolor, dignissim vitae risus vel, tristique egestas sapien.
       Vivamus blandit augue justo, id tincidunt justo luctus et. Morbi lacinia porttitor lacus, ac fermentum mauris tempus dapibus. Donec fringilla est ut ullamcorper consequat.
       Aliquam ornare efficitur ornare. Curabitur facilisis urna a congue gravida.
       Nulla accumsan non nisl sed elementum.