+++ packages/terra-core-docs/src/terra-dev-site/test/action-header/BackActionHeader.test.jsx
@@ -4,7 +4,7 @@ import ActionHeader from 'terra-action-header/lib/ActionHeader';
 
 const ActionHeaderExample = () => (
   <ActionHeader
-    title="Back Action Header"
+    text="Back Action Header"
     onBack={() => alert('You clicked back!')}
   />
 );